requirements.txt
customtkinter==5.2.0
pyserial==3.5
matplotlib==3.7.2
pandas==2.0.3